Dir refactor (#788)
* Directory layout changes

* New Makefiles

* Update import paths

* Add gtest build

* Update references in README

* Address PR feedback
