+++ packages/google-compute-engine-oslogin/selinux/oslogin.te
