+++ packages/google-compute-engine-oslogin/src/include/compat.h
