+++ packages/google-compute-engine-oslogin/src/include/oslogin_utils.h
