+++ packages/google-compute-engine-oslogin/src/nss/compat/getpwent_r.c
