+++ packages/google-compute-engine-oslogin/src/pam/pam_oslogin_admin.cc
@@ -25,8 +25,8 @@
 #include <sstream>
 #include <string>
 
-#include "../compat.h"
-#include "../utils/oslogin_utils.h"
+#include <compat.h>
+#include <oslogin_utils.h>
 
 using std::string;
